Moved simulation function and its helpers to its own file
